requirements.txt
pytest
numpy
matplotlib
pillow
